big update to actions

+++ code/patchwork/patchwork_core/src/termination/termination.rs
@@ -14,9 +14,9 @@ pub struct Termination {
     /// The type of termination of the game.
     pub termination: TerminationType,
     /// The score of player 1.
-    pub player_1_score: isize,
+    pub player_1_score: i32,
     /// The score of player 2.
-    pub player_2_score: isize,
+    pub player_2_score: i32,
 }
 impl Termination {
     /// Returns the score of the game. Positive if player 1 won, negative if player 2 won, 0 if draw. The score is calculated by taking the difference between the score of player 1 and player 2.
@@ -25,7 +25,7 @@ impl Termination {
     ///
     /// The score of the game.
     #[inline]
-    pub fn score(&self) -> isize {
+    pub fn score(&self) -> i32 {
         self.player_1_score - self.player_2_score
     }
 }
