+++ code/patchwork/patchwork_evaluator/src/win_loss_evaluator.rs
@@ -17,11 +17,11 @@ impl Default for WinLossEvaluator {
 }
 
 impl Evaluator for WinLossEvaluator {
-    fn evaluate_intermediate_node(&self, game: &Patchwork) -> isize {
+    fn evaluate_intermediate_node(&self, game: &Patchwork) -> i32 {
         self.evaluate_terminal_node(&game.random_rollout())
     }
 
-    fn evaluate_terminal_node(&self, game: &Patchwork) -> isize {
+    fn evaluate_terminal_node(&self, game: &Patchwork) -> i32 {
         match game.get_termination_result().termination {
             TerminationType::Player1Won => 1,
             TerminationType::Player2Won => -1,
